Refactors goal based methods in to new goal class and adjusts requires
